requirements.txt
# TODO define versions!
requests
python-Levenshtein
click
pyyaml